fix error UI

+++ frontend/src/services/groupService.js
@@ -1,4 +1,4 @@
-import apiClient from '../lib/api/client.js';
+import apiClient from './api/interceptors.js';
 
 /**
  * Group Service
@@ -8,63 +8,63 @@ class GroupService {
    * Create new group
    */
   async createGroup(groupData) {
-    return await apiClient.post('/user', groupData);
+    return await apiClient.post('/user/groups', groupData);
   }
 
   /**
    * Get all groups for current user
    */
   async getUserGroups() {
-    return await apiClient.get('/user');
+    return await apiClient.get('/user/groups');
   }
 
   /**
    * Get group by ID
    */
   async getGroupById(groupId) {
-    return await apiClient.get(`/user/${groupId}`);
+    return await apiClient.get(`/user/groups/${groupId}`);
   }
 
   /**
    * Update group
    */
   async updateGroup(groupId, groupData) {
-    return await apiClient.put(`/user/${groupId}`, groupData);
+    return await apiClient.put(`/user/groups/${groupId}`, groupData);
   }
 
   /**
    * Delete group
    */
   async deleteGroup(groupId) {
-    return await apiClient.delete(`/user/${groupId}`);
+    return await apiClient.delete(`/user/groups/${groupId}`);
   }
 
   /**
    * Get group members
    */
   async getGroupMembers(groupId) {
-    return await apiClient.get(`/user/${groupId}/members`);
+    return await apiClient.get(`/user/groups/${groupId}/members`);
   }
 
   /**
    * Add member to group
    */
   async addMember(groupId, memberData) {
-    return await apiClient.post(`/user/${groupId}/members`, memberData);
+    return await apiClient.post(`/user/groups/${groupId}/members`, memberData);
   }
 
   /**
    * Remove member from group
    */
   async removeMember(groupId, userId) {
-    return await apiClient.delete(`/user/${groupId}/members/${userId}`);
+    return await apiClient.delete(`/user/groups/${groupId}/members/${userId}`);
   }
 
   /**
    * Update member ownership percentage
    */
   async updateOwnership(groupId, userId, ownershipData) {
-    return await apiClient.put(`/user/${groupId}/members/${userId}/ownership`, ownershipData);
+    return await apiClient.put(`/user/groups/${groupId}/members/${userId}/ownership`, ownershipData);
   }
 }
